some Bugs and UIBugs in iPhone 4s Size fixed - crashing app after new user and skip tutorial and then add google Acount and go to shop fixed - some other bugs fixed also

+++ footbalika/Models/gameDataModel.swift
@@ -20,7 +20,7 @@ public class gameDataModel {
         let gameCharge : [gameCharge.response]
         let stadiumData : [stadiumData.response]
         let userXps : [userXps.response]
-        let onLineTime : Int?
+        let onLineTime : Float?
         let androidForceUpdateVersion : Int?
         let server_update : Int?
         let hasPrediction : Int?
